lang/codes/opsmanager/opsmaster/pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Sseve", email = "719283952@qq.com"},
]
dependencies = [
    "fastapi[all]>=0.92.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "mongoengine[all]>=0.26.0",
    "aioredis[all]>=2.0.1",
    "celery>=5.2.7",
    "gevent>=22.10.2",
    "grpcio>=1.51.1",
    "grpcio-tools>=1.51.1",
    "redis>=4.5.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}
